调整上传数据为 i32

+++ src/models.rs
@@ -22,8 +22,8 @@ pub struct Target {
 #[derive(Serialize, Deserialize, Debug)]
 pub struct PingData {
     pub ipv6: bool,
-    pub min: f32,
-    pub jitter: f32,
+    pub min: i32,
+    pub jitter: i32,
     pub failed: i32,
 }
